Partial refactor media retention policy
