+++ genesyscloud/data_source_genesyscloud_routing_email_domain_test.go
@@ -2,43 +2,12 @@ package genesyscloud
 
 import (
 	"fmt"
-	"log"
 	"math/rand"
 	"strconv"
-	"strings"
 	"testing"
-	"time"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/mypurecloud/platform-client-sdk-go/v105/platformclientv2"
 )
-
-func cleanupRoutingEmailDomains() {
-	routingAPI := platformclientv2.NewRoutingApiWithConfig(sdkConfig)
-
-	for pageNum := 1; ; pageNum++ {
-		const pageSize = 100
-		routingEmailDomains, _, getErr := routingAPI.GetRoutingEmailDomains(pageNum, pageSize, false, "")
-		if getErr != nil {
-			return
-		}
-
-		if routingEmailDomains.Entities == nil || len(*routingEmailDomains.Entities) == 0 {
-			return
-		}
-
-		for _, routingEmailDomain := range *routingEmailDomains.Entities {
-			if routingEmailDomain.Id != nil && strings.HasPrefix(*routingEmailDomain.Id, "terraform") {
-				_, err := routingAPI.DeleteRoutingEmailDomain(*routingEmailDomain.Id)
-				if err != nil {
-					log.Printf("Failed to delete routing email domain %s: %s", *routingEmailDomain.Id, err)
-					continue
-				}
-				time.Sleep(5 * time.Second)
-			}
-		}
-	}
-}
 
 func TestAccDataSourceRoutingEmailDomain(t *testing.T) {
 	var (
@@ -51,7 +20,7 @@ func TestAccDataSourceRoutingEmailDomain(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	cleanupRoutingEmailDomains()
+	CleanupRoutingEmailDomains()
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { TestAccPreCheck(t) },
