+++ genesyscloud/recording_media_retention_policy/data_source_genesyscloud_recording_media_retention_policy_test.go
@@ -1,10 +1,12 @@
-package genesyscloud
+package recording_media_retention_policy
 
 import (
 	"fmt"
 	"math/rand"
 	"strconv"
 	"testing"
+
+	gcloud "terraform-provider-genesyscloud/genesyscloud"
 
 	"github.com/google/uuid"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
@@ -125,15 +127,15 @@ func TestAccDataSourceRecordingMediaRetentionPolicy(t *testing.T) {
 		domainId  = "terraform" + strconv.Itoa(rand.Intn(1000)) + ".com"
 	)
 
-	_, err := AuthorizeSdk()
+	_, err := gcloud.AuthorizeSdk()
 	if err != nil {
 		t.Fatal(err)
 	}
-	cleanupRoutingEmailDomains()
+	gcloud.CleanupRoutingEmailDomains()
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { TestAccPreCheck(t) },
-		ProviderFactories: GetProviderFactories(providerResources, providerDataSources),
+		PreCheck:          func() { gcloud.TestAccPreCheck(t) },
+		ProviderFactories: gcloud.GetProviderFactories(providerResources, providerDataSources),
 		Steps: []resource.TestStep{
 			{
 				Config: generateRoutingEmailDomainResource(
