+++ genesyscloud/recording_media_retention_policy/genesyscloud_recording_media_retention_policy_init_test.go
@@ -1,0 +1,1 @@
+package recording_media_retention_policy
