+++ genesyscloud/recording_media_retention_policy/resource_genesyscloud_recording_media_retention_policy_utils.go
@@ -1,21 +1,12 @@
-package genesyscloud
+package recording_media_retention_policy
 
 import (
-	"context"
 	"fmt"
 	"log"
 	"reflect"
 	"time"
 
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
-
-	"terraform-provider-genesyscloud/genesyscloud/consistency_checker"
-
-	resourceExporter "terraform-provider-genesyscloud/genesyscloud/resource_exporter"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/mypurecloud/platform-client-sdk-go/v105/platformclientv2"
 )
 
@@ -57,938 +48,7 @@ type VisibilityConditionStruct struct {
 	Predicates         []string
 }
 
-var (
-	mediaPolicies = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"call_policy": {
-				Description: "Conditions and actions for calls",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        callMediaPolicy,
-			},
-			"chat_policy": {
-				Description: "Conditions and actions for calls",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        chatMediaPolicy,
-			},
-			"email_policy": {
-				Description: "Conditions and actions for calls",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        emailMediaPolicy,
-			},
-			"message_policy": {
-				Description: "Conditions and actions for calls",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        messageMediaPolicy,
-			},
-		},
-	}
-
-	chatMediaPolicy = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"actions": {
-				Description: "Actions applied when specified conditions are met",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        policyActions,
-			},
-			"conditions": {
-				Description: "Conditions for when actions should be applied",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        chatMediaPolicyConditions,
-			},
-		},
-	}
-
-	callMediaPolicy = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"actions": {
-				Description: "Actions applied when specified conditions are met",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        policyActions,
-			},
-			"conditions": {
-				Description: "Conditions for when actions should be applied",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        callMediaPolicyConditions,
-			},
-		},
-	}
-
-	emailMediaPolicy = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"actions": {
-				Description: "Actions applied when specified conditions are met",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        policyActions,
-			},
-			"conditions": {
-				Description: "Conditions for when actions should be applied",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        emailMediaPolicyConditions,
-			},
-		},
-	}
-
-	messageMediaPolicy = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"actions": {
-				Description: "Actions applied when specified conditions are met",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        policyActions,
-			},
-			"conditions": {
-				Description: "Conditions for when actions should be applied",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        messageMediaPolicyConditions,
-			},
-		},
-	}
-
-	policyActions = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"retain_recording": {
-				Description: "true to retain the recording associated with the conversation.",
-				Type:        schema.TypeBool,
-				Optional:    true,
-			},
-			"delete_recording": {
-				Description: "true to delete the recording associated with the conversation. If retainRecording = true, this will be ignored.",
-				Type:        schema.TypeBool,
-				Optional:    true,
-			},
-			"always_delete": {
-				Description: "true to delete the recording associated with the conversation regardless of the values of retainRecording or deleteRecording.",
-				Type:        schema.TypeBool,
-				Optional:    true,
-			},
-			"assign_evaluations": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        evaluationAssignment,
-			},
-			"assign_metered_evaluations": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        meteredEvaluationAssignment,
-			},
-			"assign_metered_assignment_by_agent": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        meteredAssignmentByAgent,
-			},
-			"assign_calibrations": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        calibrationAssignment,
-			},
-			"assign_surveys": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        surveyAssignment,
-			},
-			"retention_duration": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        retentionDuration,
-			},
-			"initiate_screen_recording": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        initiateScreenRecording,
-			},
-			"media_transcriptions": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        mediaTranscription,
-			},
-			"integration_export": {
-				Description: "Policy action for exporting recordings using an integration to 3rd party s3.",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        integrationExport,
-			},
-		},
-	}
-
-	evaluationAssignment = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"evaluation_form_id": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"user_id": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-		},
-	}
-
-	meteredEvaluationAssignment = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"evaluator_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"max_number_evaluations": {
-				Description: "",
-				Type:        schema.TypeInt,
-				Optional:    true,
-			},
-			"evaluation_form_id": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"assign_to_active_user": {
-				Description: "",
-				Type:        schema.TypeBool,
-				Optional:    true,
-			},
-			"time_interval": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        timeInterval,
-			},
-		},
-	}
-
-	meteredAssignmentByAgent = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"evaluator_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"max_number_evaluations": {
-				Description: "",
-				Type:        schema.TypeInt,
-				Optional:    true,
-			},
-			"evaluation_form_id": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"time_interval": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        timeInterval,
-			},
-			"time_zone": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-		},
-	}
-
-	calibrationAssignment = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"calibrator_id": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"evaluator_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"evaluation_form_id": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"expert_evaluator_id": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-		},
-	}
-
-	surveyAssignment = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"survey_form_name": {
-				Description: "The survey form used for this survey.",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"flow_id": {
-				Description: "The UUID reference to the flow associated with this survey.",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"invite_time_interval": {
-				Description: "An ISO 8601 repeated interval consisting of the number of repetitions, the start datetime, and the interval (e.g. R2/2018-03-01T13:00:00Z/P1M10DT2H30M). Total duration must not exceed 90 days.",
-				Type:        schema.TypeString,
-				Default:     "R1/P0M",
-				Optional:    true,
-			},
-			"sending_user": {
-				Description: "User together with sendingDomain used to send email, null to use no-reply",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"sending_domain": {
-				Description: "Validated email domain, required",
-				Type:        schema.TypeString,
-				Required:    true,
-			},
-		},
-	}
-
-	retentionDuration = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"archive_retention": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        archiveRetention,
-			},
-			"delete_retention": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        deleteRetention,
-			},
-		},
-	}
-
-	initiateScreenRecording = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"record_acw": {
-				Description: "",
-				Type:        schema.TypeBool,
-				Optional:    true,
-			},
-			"archive_retention": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        archiveRetention,
-			},
-			"delete_retention": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        deleteRetention,
-			},
-		},
-	}
-
-	mediaTranscription = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"display_name": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"transcription_provider": {
-				Description:  "",
-				Type:         schema.TypeString,
-				Optional:     true,
-				ValidateFunc: validation.StringInSlice([]string{"VOCI", "CALLJOURNEY"}, false),
-			},
-			"integration_id": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-		},
-	}
-
-	integrationExport = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"integration_id": {
-				Description: "The aws-s3-recording-bulk-actions-integration that the policy uses for exports.",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"should_export_screen_recordings": {
-				Description: "True if the policy should export screen recordings in addition to the other conversation media.",
-				Type:        schema.TypeBool,
-				Optional:    true,
-				Default:     true,
-			},
-		},
-	}
-
-	timeInterval = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"months": {
-				Description: "",
-				Type:        schema.TypeInt,
-				Optional:    true,
-			},
-			"weeks": {
-				Description: "",
-				Type:        schema.TypeInt,
-				Optional:    true,
-			},
-			"days": {
-				Description: "",
-				Type:        schema.TypeInt,
-				Optional:    true,
-			},
-			"hours": {
-				Description: "",
-				Type:        schema.TypeInt,
-				Optional:    true,
-			},
-		},
-	}
-
-	policyErrors = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"policy_error_messages": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        policyErrorMessage,
-			},
-		},
-	}
-
-	policyErrorMessage = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"status_code": {
-				Description: "",
-				Type:        schema.TypeInt,
-				Optional:    true,
-			},
-			"user_message": {
-				Description: "",
-				Type:        schema.TypeMap,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-				Optional:    true,
-			},
-			"user_params_message": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"error_code": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"correlation_id": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"user_params": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        userParam,
-			},
-			"insert_date": {
-				Description: "Date time is represented as an ISO-8601 string. For example: yyyy-MM-ddTHH:mm:ss[.mmm]Z",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-		},
-	}
-
-	userParam = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"key": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"value": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-		},
-	}
-
-	archiveRetention = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"days": {
-				Description: "",
-				Type:        schema.TypeInt,
-				Optional:    true,
-			},
-			"storage_medium": {
-				Description:  "",
-				Type:         schema.TypeString,
-				Optional:     true,
-				ValidateFunc: validation.StringInSlice([]string{"CLOUDARCHIVE"}, false)},
-		},
-	}
-
-	deleteRetention = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"days": {
-				Description: "",
-				Type:        schema.TypeInt,
-				Optional:    true,
-			},
-		},
-	}
-
-	timeSlot = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"start_time": {
-				Description: "start time in xx:xx:xx.xxx format",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"stop_time": {
-				Description: "stop time in xx:xx:xx.xxx format",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"day": {
-				Description: "Day for this time slot, Monday = 1 ... Sunday = 7",
-				Type:        schema.TypeInt,
-				Optional:    true,
-			},
-		},
-	}
-
-	timeAllowed = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"time_slots": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        timeSlot,
-			},
-			"time_zone_id": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"empty": {
-				Description: "",
-				Type:        schema.TypeBool,
-				Optional:    true,
-			},
-		},
-	}
-
-	durationCondition = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"duration_target": {
-				Description:  "",
-				Type:         schema.TypeString,
-				Optional:     true,
-				ValidateFunc: validation.StringInSlice([]string{"DURATION", "DURATION_RANGE"}, false),
-			},
-			"duration_operator": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"duration_range": {
-				Description: "",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"duration_mode": {
-				Description:  "",
-				Type:         schema.TypeString,
-				Optional:     true,
-				ValidateFunc: validation.StringInSlice([]string{"Between", "Over", "Under"}, false),
-			},
-		},
-	}
-
-	callMediaPolicyConditions = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"for_user_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"date_ranges": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Default:     nil,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"for_queue_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"wrapup_code_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"language_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"time_allowed": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        timeAllowed,
-			},
-			"directions": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Default:     nil,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"duration": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        durationCondition,
-			},
-		},
-	}
-
-	chatMediaPolicyConditions = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"for_user_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"date_ranges": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Default:     nil,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"for_queue_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"wrapup_code_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"language_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"time_allowed": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        timeAllowed,
-			},
-			"duration": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        durationCondition,
-			},
-		},
-	}
-
-	emailMediaPolicyConditions = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"for_user_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"date_ranges": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Default:     nil,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"for_queue_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"wrapup_code_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"language_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"time_allowed": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        timeAllowed,
-			},
-		},
-	}
-
-	messageMediaPolicyConditions = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"for_user_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"date_ranges": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Default:     nil,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"for_queue_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"wrapup_code_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"language_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"time_allowed": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        timeAllowed,
-			},
-		},
-	}
-
-	policyConditions = &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"for_user_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"directions": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Default:     nil,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"date_ranges": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Default:     nil,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"media_types": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"for_queue_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"duration": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        durationCondition,
-			},
-			"wrapup_code_ids": {
-				Description: "",
-				Type:        schema.TypeList,
-				Optional:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-			"time_allowed": {
-				Description: "",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        timeAllowed,
-			},
-		},
-	}
-
-	qualityAPI = platformclientv2.NewQualityApi()
-)
-
-func ResourceMediaRetentionPolicy() *schema.Resource {
-	return &schema.Resource{
-		Description:   "Genesys Cloud Media Retention Policies",
-		CreateContext: CreateWithPooledClient(createMediaRetentionPolicy),
-		ReadContext:   ReadWithPooledClient(readMediaRetentionPolicy),
-		UpdateContext: UpdateWithPooledClient(updateMediaRetentionPolicy),
-		DeleteContext: DeleteWithPooledClient(deleteMediaRetentionPolicy),
-		Importer: &schema.ResourceImporter{
-			StateContext: schema.ImportStatePassthroughContext,
-		},
-		SchemaVersion: 1,
-		Schema: map[string]*schema.Schema{
-			"name": {
-				Description: "The policy name. Changing the policy_name attribute will cause the recording_media_retention_policy to be dropped and recreated with a new ID.",
-				Type:        schema.TypeString,
-				Required:    true,
-				ForceNew:    true,
-			},
-			"order": {
-				Description: "The ordinal number for the policy",
-				Type:        schema.TypeInt,
-				Optional:    true,
-			},
-			"description": {
-				Description: "The description for the policy",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"enabled": {
-				Description: "The policy will be enabled if true, otherwise it will be disabled",
-				Type:        schema.TypeBool,
-				Optional:    true,
-			},
-			"media_policies": {
-				Description: "Conditions and actions per media type",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        mediaPolicies,
-			},
-			"conditions": {
-				Description: "Conditions",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        policyConditions,
-			},
-			"actions": {
-				Description: "Actions",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        policyActions,
-			},
-			"policy_errors": {
-				Description: "A list of errors in the policy configuration",
-				Type:        schema.TypeList,
-				MaxItems:    1,
-				Optional:    true,
-				Elem:        policyErrors,
-			},
-		},
-	}
-}
-
-func getAllMediaRetentionPolicies(_ context.Context, clientConfig *platformclientv2.Configuration) (resourceExporter.ResourceIDMetaMap, diag.Diagnostics) {
-	resources := make(resourceExporter.ResourceIDMetaMap)
-	recordingAPI := platformclientv2.NewRecordingApiWithConfig(clientConfig)
-
-	for pageNum := 1; ; pageNum++ {
-		const pageSize = 100
-		retentionPolicies, _, getErr := recordingAPI.GetRecordingMediaretentionpolicies(pageSize, pageNum, "", []string{}, "", "", "", true, false, false, 365)
-
-		if getErr != nil {
-			return nil, diag.Errorf("Failed to get page of media retention policies %v", getErr)
-		}
-
-		if retentionPolicies.Entities == nil || len(*retentionPolicies.Entities) == 0 {
-			break
-		}
-
-		for _, retentionPolicy := range *retentionPolicies.Entities {
-			resources[*retentionPolicy.Id] = &resourceExporter.ResourceMeta{Name: *retentionPolicy.Name}
-		}
-	}
-
-	return resources, nil
-}
+var qualityAPI = platformclientv2.NewQualityApi()
 
 func buildEvaluationAssignments(evaluations []interface{}) *[]platformclientv2.Evaluationassignment {
 	assignEvaluations := make([]platformclientv2.Evaluationassignment, 0)
@@ -1667,7 +727,7 @@ func flattenIntegrationExport(integrationExport *platformclientv2.Integrationexp
 	return []interface{}{integrationExportMap}
 }
 
-func buildPolicyActions(actions []interface{}) *platformclientv2.Policyactions {
+func buildPolicyActionsFromMediaPolicy(actions []interface{}) *platformclientv2.Policyactions {
 	if actions == nil || len(actions) <= 0 {
 		return nil
 	}
@@ -2371,7 +1431,7 @@ func buildCallMediaPolicy(callMediaPolicy []interface{}) *platformclientv2.Callm
 		return nil
 	}
 	return &platformclientv2.Callmediapolicy{
-		Actions:    buildPolicyActions(policyMap["actions"].([]interface{})),
+		Actions:    buildPolicyActionsFromMediaPolicy(policyMap["actions"].([]interface{})),
 		Conditions: buildCallMediaPolicyConditions(policyMap["conditions"].([]interface{})),
 	}
 }
@@ -2403,7 +1463,7 @@ func buildChatMediaPolicy(chatMediaPolicy []interface{}) *platformclientv2.Chatm
 	}
 
 	return &platformclientv2.Chatmediapolicy{
-		Actions:    buildPolicyActions(policyMap["actions"].([]interface{})),
+		Actions:    buildPolicyActionsFromMediaPolicy(policyMap["actions"].([]interface{})),
 		Conditions: buildChatMediaPolicyConditions(policyMap["conditions"].([]interface{})),
 	}
 }
@@ -2435,7 +1495,7 @@ func buildEmailMediaPolicy(emailMediaPolicy []interface{}) *platformclientv2.Ema
 	}
 
 	return &platformclientv2.Emailmediapolicy{
-		Actions:    buildPolicyActions(policyMap["actions"].([]interface{})),
+		Actions:    buildPolicyActionsFromMediaPolicy(policyMap["actions"].([]interface{})),
 		Conditions: buildEmailMediaPolicyConditions(policyMap["conditions"].([]interface{})),
 	}
 }
@@ -2467,7 +1527,7 @@ func buildMessageMediaPolicy(messageMediaPolicy []interface{}) *platformclientv2
 	}
 
 	return &platformclientv2.Messagemediapolicy{
-		Actions:    buildPolicyActions(policyMap["actions"].([]interface{})),
+		Actions:    buildPolicyActionsFromMediaPolicy(policyMap["actions"].([]interface{})),
 		Conditions: buildMessageMediaPolicyConditions(policyMap["conditions"].([]interface{})),
 	}
 }
@@ -2660,7 +1720,7 @@ func flattenConditions(conditions *platformclientv2.Policyconditions) []interfac
 	return []interface{}{conditionsMap}
 }
 
-func buildPolicyActions2(d *schema.ResourceData) *platformclientv2.Policyactions {
+func buildPolicyActionsFromResource(d *schema.ResourceData) *platformclientv2.Policyactions {
 
 	if actions, ok := d.Get("actions").([]interface{}); ok && len(actions) > 0 {
 		actionsMap, ok := actions[0].(map[string]interface{})
@@ -2832,244 +1892,4 @@ func flattenPolicyErrors(policyErrors *platformclientv2.Policyerrors) []interfac
 	}
 
 	return []interface{}{policyErrorsMap}
-}
-
-func readMediaRetentionPolicy(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	sdkConfig := meta.(*ProviderMeta).ClientConfig
-	recordingAPI := platformclientv2.NewRecordingApiWithConfig(sdkConfig)
-
-	log.Printf("Reading media retention policy %s", d.Id())
-
-	return WithRetriesForRead(ctx, d, func() *retry.RetryError {
-		retentionPolicy, resp, getErr := recordingAPI.GetRecordingMediaretentionpolicy(d.Id())
-		if getErr != nil {
-			if IsStatus404(resp) {
-				return retry.RetryableError(fmt.Errorf("failed to read media retention policy %s: %s", d.Id(), getErr))
-			}
-			return retry.NonRetryableError(fmt.Errorf("failed to read media retention policy %s: %s", d.Id(), getErr))
-		}
-
-		cc := consistency_checker.NewConsistencyCheck(ctx, d, meta, ResourceSurveyForm())
-		if retentionPolicy.Name != nil {
-			d.Set("name", *retentionPolicy.Name)
-		}
-		if retentionPolicy.Order != nil {
-			d.Set("order", *retentionPolicy.Order)
-		}
-		if retentionPolicy.Description != nil {
-			d.Set("description", *retentionPolicy.Description)
-		}
-		if retentionPolicy.Enabled != nil {
-			d.Set("enabled", *retentionPolicy.Enabled)
-		}
-		if retentionPolicy.MediaPolicies != nil {
-			d.Set("media_policies", flattenMediaPolicies(retentionPolicy.MediaPolicies))
-		}
-		if retentionPolicy.Conditions != nil {
-			d.Set("conditions", flattenConditions(retentionPolicy.Conditions))
-		}
-		if retentionPolicy.Actions != nil {
-			d.Set("actions", flattenPolicyActions(retentionPolicy.Actions))
-		}
-		if retentionPolicy.PolicyErrors != nil {
-			d.Set("policy_errors", flattenPolicyErrors(retentionPolicy.PolicyErrors))
-		}
-
-		return cc.CheckState()
-	})
-}
-
-func createMediaRetentionPolicy(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	name := d.Get("name").(string)
-	order := d.Get("order").(int)
-	description := d.Get("description").(string)
-	enabled := d.Get("enabled").(bool)
-	mediaPolicies := buildMediaPolicies(d)
-	conditions := buildConditions(d)
-	actions := buildPolicyActions2(d)
-	policyErrors := buildPolicyErrors(d)
-	sdkConfig := meta.(*ProviderMeta).ClientConfig
-	recordingAPI := platformclientv2.NewRecordingApiWithConfig(sdkConfig)
-
-	reqBody := platformclientv2.Policycreate{
-		Name:          &name,
-		Order:         &order,
-		Description:   &description,
-		Enabled:       &enabled,
-		MediaPolicies: mediaPolicies,
-		Conditions:    conditions,
-		Actions:       actions,
-		PolicyErrors:  policyErrors,
-	}
-
-	log.Printf("Creating media retention policy %s", name)
-	policy, apiResponse, err := recordingAPI.PostRecordingMediaretentionpolicies(reqBody)
-	log.Printf("Media retention policy creation status %#v", apiResponse.Status)
-
-	if err != nil {
-		return diag.Errorf("Failed to create media retention policy %s: %s", name, err)
-	}
-
-	// Make sure form is properly created
-	policyId := policy.Id
-	d.SetId(*policyId)
-	log.Printf("Created media retention policy %s %s", name, *policy.Id)
-	return readMediaRetentionPolicy(ctx, d, meta)
-}
-
-func updateMediaRetentionPolicy(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	name := d.Get("name").(string)
-	order := d.Get("order").(int)
-	description := d.Get("description").(string)
-	enabled := d.Get("enabled").(bool)
-
-	mediaPolicies := buildMediaPolicies(d)
-	conditions := buildConditions(d)
-	actions := buildPolicyActions2(d)
-	policyErrors := buildPolicyErrors(d)
-
-	sdkConfig := meta.(*ProviderMeta).ClientConfig
-	recordingAPI := platformclientv2.NewRecordingApiWithConfig(sdkConfig)
-
-	reqBody := platformclientv2.Policy{
-		Name:          &name,
-		Order:         &order,
-		Description:   &description,
-		Enabled:       &enabled,
-		MediaPolicies: mediaPolicies,
-		Conditions:    conditions,
-		Actions:       actions,
-		PolicyErrors:  policyErrors,
-	}
-
-	log.Printf("Updating media retention policy %s", name)
-	policy, _, err := recordingAPI.PutRecordingMediaretentionpolicy(d.Id(), reqBody)
-	if err != nil {
-		return diag.Errorf("Failed to update media retention policy %s: %s", name, err)
-	}
-
-	log.Printf("Updated media retention policy %s %s", name, *policy.Id)
-	return readMediaRetentionPolicy(ctx, d, meta)
-}
-
-func MediaRetentionPolicyExporter() *resourceExporter.ResourceExporter {
-	return &resourceExporter.ResourceExporter{
-		GetResourcesFunc: GetAllWithPooledClient(getAllMediaRetentionPolicies),
-		RefAttrs: map[string]*resourceExporter.RefAttrSettings{
-			"media_policies.chat_policy.conditions.for_queue_ids":                                         {RefType: "genesyscloud_routing_queue", AltValues: []string{"*"}},
-			"media_policies.call_policy.conditions.for_queue_ids":                                         {RefType: "genesyscloud_routing_queue", AltValues: []string{"*"}},
-			"media_policies.message_policy.conditions.for_queue_ids":                                      {RefType: "genesyscloud_routing_queue", AltValues: []string{"*"}},
-			"media_policies.email_policy.conditions.for_queue_ids":                                        {RefType: "genesyscloud_routing_queue", AltValues: []string{"*"}},
-			"conditions.for_queue_ids":                                                                    {RefType: "genesyscloud_routing_queue", AltValues: []string{"*"}},
-			"media_policies.call_policy.conditions.for_user_ids":                                          {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.chat_policy.conditions.for_user_ids":                                          {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.email_policy.conditions.for_user_ids":                                         {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.message_policy.conditions.for_user_ids":                                       {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"conditions.for_user_ids":                                                                     {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.call_policy.actions.assign_evaluations.evaluation_form_id":                    {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.call_policy.actions.assign_calibrations.evaluation_form_id":                   {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.call_policy.actions.assign_metered_evaluations.evaluation_form_id":            {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.call_policy.actions.assign_metered_assignment_by_agent.evaluation_form_id":    {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.chat_policy.actions.assign_evaluations.evaluation_form_id":                    {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.chat_policy.actions.assign_calibrations.evaluation_form_id":                   {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.chat_policy.actions.assign_metered_evaluations.evaluation_form_id":            {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.chat_policy.actions.assign_metered_assignment_by_agent.evaluation_form_id":    {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.message_policy.actions.assign_evaluations.evaluation_form_id":                 {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.message_policy.actions.assign_calibrations.evaluation_form_id":                {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.message_policy.actions.assign_metered_evaluations.evaluation_form_id":         {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.message_policy.actions.assign_metered_assignment_by_agent.evaluation_form_id": {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.email_policy.actions.assign_evaluations.evaluation_form_id":                   {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.email_policy.actions.assign_calibrations.evaluation_form_id":                  {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.email_policy.actions.assign_metered_evaluations.evaluation_form_id":           {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.email_policy.actions.assign_metered_assignment_by_agent.evaluation_form_id":   {RefType: "genesyscloud_quality_forms_evaluation"},
-			"actions.assign_evaluations.evaluation_form_id":                                               {RefType: "genesyscloud_quality_forms_evaluation"},
-			"actions.assign_calibrations.evaluation_form_id":                                              {RefType: "genesyscloud_quality_forms_evaluation"},
-			"actions.assign_metered_evaluations.evaluation_form_id":                                       {RefType: "genesyscloud_quality_forms_evaluation"},
-			"actions.assign_metered_assignment_by_agent.evaluation_form_id":                               {RefType: "genesyscloud_quality_forms_evaluation"},
-			"media_policies.call_policy.actions.assign_evaluations.evaluator_ids":                         {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.call_policy.actions.assign_calibrations.evaluator_ids":                        {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.call_policy.actions.assign_metered_evaluations.evaluator_ids":                 {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.call_policy.actions.assign_metered_assignment_by_agent.evaluator_ids":         {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.chat_policy.actions.assign_evaluations.evaluator_ids":                         {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.chat_policy.actions.assign_calibrations.evaluator_ids":                        {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.chat_policy.actions.assign_metered_evaluations.evaluator_ids":                 {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.chat_policy.actions.assign_metered_assignment_by_agent.evaluator_ids":         {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.message_policy.actions.assign_evaluations.evaluator_ids":                      {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.message_policy.actions.assign_calibrations.evaluator_ids":                     {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.message_policy.actions.assign_metered_evaluations.evaluator_ids":              {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.message_policy.actions.assign_metered_assignment_by_agent.evaluator_ids":      {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.email_policy.actions.assign_evaluations.evaluator_ids":                        {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.email_policy.actions.assign_calibrations.evaluator_ids":                       {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.email_policy.actions.assign_metered_evaluations.evaluator_ids":                {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.email_policy.actions.assign_metered_assignment_by_agent.evaluator_ids":        {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"actions.assign_evaluations.evaluator_ids":                                                    {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"actions.assign_calibrations.evaluator_ids":                                                   {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"actions.assign_metered_evaluations.evaluator_ids":                                            {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"actions.assign_metered_assignment_by_agent.evaluator_ids":                                    {RefType: "genesyscloud_user", AltValues: []string{"*"}},
-			"media_policies.call_policy.actions.assign_calibrations.calibrator_id":                        {RefType: "genesyscloud_user"},
-			"media_policies.chat_policy.actions.assign_calibrations.calibrator_id":                        {RefType: "genesyscloud_user"},
-			"media_policies.message_policy.actions.assign_calibrations.calibrator_id":                     {RefType: "genesyscloud_user"},
-			"media_policies.email_policy.actions.assign_calibrations.calibrator_id":                       {RefType: "genesyscloud_user"},
-			"media_policies.call_policy.actions.assign_calibrations.expert_evaluator_id":                  {RefType: "genesyscloud_user"},
-			"media_policies.chat_policy.actions.assign_calibrations.expert_evaluator_id":                  {RefType: "genesyscloud_user"},
-			"media_policies.message_policy.actions.assign_calibrations.expert_evaluator_id":               {RefType: "genesyscloud_user"},
-			"media_policies.email_policy.actions.assign_calibrations.expert_evaluator_id":                 {RefType: "genesyscloud_user"},
-			"actions.assign_calibrations.expert_evaluator_id":                                             {RefType: "genesyscloud_user"},
-			"media_policies.call_policy.conditions.language_ids":                                          {RefType: "genesyscloud_routing_language", AltValues: []string{"*"}},
-			"media_policies.chat_policy.conditions.language_ids":                                          {RefType: "genesyscloud_routing_language", AltValues: []string{"*"}},
-			"media_policies.message_policy.conditions.language_ids":                                       {RefType: "genesyscloud_routing_language", AltValues: []string{"*"}},
-			"media_policies.email_policy.conditions.language_ids":                                         {RefType: "genesyscloud_routing_language", AltValues: []string{"*"}},
-			"media_policies.call_policy.conditions.wrapup_code_ids":                                       {RefType: "genesyscloud_routing_wrapupcode", AltValues: []string{"*"}},
-			"media_policies.chat_policy.conditions.wrapup_code_ids":                                       {RefType: "genesyscloud_routing_wrapupcode", AltValues: []string{"*"}},
-			"media_policies.message_policy.conditions.wrapup_code_ids":                                    {RefType: "genesyscloud_routing_wrapupcode", AltValues: []string{"*"}},
-			"media_policies.email_policy.conditions.wrapup_code_ids":                                      {RefType: "genesyscloud_routing_wrapupcode", AltValues: []string{"*"}},
-			"conditions.wrapup_code_ids":                                                                  {RefType: "genesyscloud_routing_wrapupcode", AltValues: []string{"*"}},
-			"media_policies.call_policy.actions.integration_export.integration_id":                        {RefType: "genesyscloud_integration"},
-			"media_policies.chat_policy.actions.integration_export.integration_id":                        {RefType: "genesyscloud_integration"},
-			"media_policies.message_policy.actions.integration_export.integration_id":                     {RefType: "genesyscloud_integration"},
-			"media_policies.email_policy.actions.integration_export.integration_id":                       {RefType: "genesyscloud_integration"},
-			"actions.media_transcriptions.integration_id":                                                 {RefType: "genesyscloud_integration"},
-			"media_policies.call_policy.actions.assign_surveys.flow_id":                                   {RefType: "genesyscloud_flow"},
-			"media_policies.chat_policy.actions.assign_surveys.flow_id":                                   {RefType: "genesyscloud_flow"},
-			"media_policies.message_policy.actions.assign_surveys.flow_id":                                {RefType: "genesyscloud_flow"},
-			"media_policies.email_policy.actions.assign_surveys.flow_id":                                  {RefType: "genesyscloud_flow"},
-			"actions.assign_surveys.flow_id":                                                              {RefType: "genesyscloud_flow"},
-			"media_policies.call_policy.actions.assign_evaluations.user_id":                               {RefType: "genesyscloud_user"},
-			"media_policies.chat_policy.actions.assign_evaluations.user_id":                               {RefType: "genesyscloud_user"},
-			"media_policies.message_policy.actions.assign_evaluations.user_id":                            {RefType: "genesyscloud_user"},
-			"media_policies.email_policy.actions.assign_evaluations.user_id":                              {RefType: "genesyscloud_user"},
-			"actions.assign_evaluations.user_id":                                                          {RefType: "genesyscloud_user"},
-		},
-		AllowZeroValues: []string{"order"},
-		RemoveIfMissing: map[string][]string{
-			"":               {"conditions", "actions"},
-			"media_policies": {"call_policy", "chat_policy", "message_policy", "email_policy"},
-		},
-	}
-}
-
-func deleteMediaRetentionPolicy(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	name := d.Get("name").(string)
-
-	sdkConfig := meta.(*ProviderMeta).ClientConfig
-	recordingAPI := platformclientv2.NewRecordingApiWithConfig(sdkConfig)
-
-	log.Printf("Deleting media retention policy %s", name)
-	if _, err := recordingAPI.DeleteRecordingMediaretentionpolicy(d.Id()); err != nil {
-		return diag.Errorf("Failed to delete media retention policy %s: %s", name, err)
-	}
-
-	return WithRetries(ctx, 30*time.Second, func() *retry.RetryError {
-		_, resp, err := recordingAPI.GetRecordingMediaretentionpolicy(d.Id())
-		if err != nil {
-			if IsStatus404(resp) {
-				// media retention policy deleted
-				log.Printf("Deleted media retention policy %s", d.Id())
-				return nil
-			}
-			return retry.NonRetryableError(fmt.Errorf("error deleting media retention policy %s: %s", d.Id(), err))
-		}
-
-		return retry.RetryableError(fmt.Errorf("media retention policy %s still exists", d.Id()))
-	})
 }
