+++ genesyscloud/resource_genesyscloud_routing_email_route_test.go
@@ -43,7 +43,7 @@ func TestAccResourceRoutingEmailRoute(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	cleanupRoutingEmailDomains()
+	CleanupRoutingEmailDomains()
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { TestAccPreCheck(t) },
